# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/sop-versions": {
      "get": {
        "operationId": "SopController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      },
      "post": {
        "operationId": "SopController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      }
    },
    "/api/v1/sop-versions/{versionId}": {
      "get": {
        "operationId": "SopController_detail",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      },
      "patch": {
        "operationId": "SopController_update",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSopVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      }
    },
    "/api/v1/sop-versions/{versionId}/validate": {
      "post": {
        "operationId": "SopController_validate",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      }
    },
    "/api/v1/sop-versions/{versionId}/publish": {
      "post": {
        "operationId": "SopController_publish",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishSopVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      }
    },
    "/api/v1/sop-versions/{versionId}/material-backfill/preview": {
      "post": {
        "operationId": "SopController_previewMaterialBackfill",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewSopMaterialBackfillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      }
    },
    "/api/v1/sop-versions/{versionId}/material-backfill/apply": {
      "post": {
        "operationId": "SopController_applyMaterialBackfill",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplySopMaterialBackfillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "sop-versions"
        ]
      }
    },
...
  },
  "info": {
    "title": "DSE升学服务CRM API",
    "description": "S1认证、权限、账号、审计与任务监督闭环接口",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dse_session"
      }
    },
    "schemas": {
...
      "SopTaskTemplateInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150
          },
          "description": {
            "type": "object",
            "maxLength": 2000,
            "nullable": true
          },
          "completionCriteria": {
            "type": "object",
            "maxLength": 2000,
            "nullable": true
          },
          "completionWindowHours": {
            "type": "number",
            "minimum": 1
          },
          "isBlocking": {
            "type": "object",
            "default": true,
            "description": "是否阻止所属阶段自动完成"
          }
        },
        "required": [
          "name",
          "completionWindowHours",
          "isBlocking"
        ]
      },
      "SopMaterialTemplateInputDto": {
        "type": "object",
        "properties": {
          "templateKey": {
            "type": "string",
            "format": "uuid",
            "description": "跨SOP版本保持不变的资料模板标识"
          },
          "materialTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 150
          },
          "requirement": {
            "type": "object",
            "maxLength": 2000,
            "nullable": true
          },
          "requirementKind": {
            "type": "string",
            "enum": [
              "REQUIRED",
              "CONDITIONAL",
              "OPTIONAL"
            ]
          },
          "deadlineRule": {
            "type": "string",
            "enum": [
              "ACTIVATION_OFFSET",
              "STAGE_OFFSET",
              "FIXED_DATE"
            ]
          },
          "deadlineOffsetDays": {
            "type": "object",
            "minimum": 0,
            "nullable": true
          },
          "fixedDueAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "conditionRule": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "materialTypeId",
          "title",
          "requirementKind",
          "deadlineRule"
        ]
      },
      "SopStageTemplateInputDto": {
        "type": "object",
        "properties": {
          "stageCode": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "object",
            "maxLength": 1000,
            "nullable": true
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SopTaskTemplateInputDto"
            }
          },
          "materials": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SopMaterialTemplateInputDto"
            }
          }
        },
        "required": [
          "stageCode",
          "name",
          "tasks"
        ]
      },
      "UpdateSopVersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1,
            "description": "乐观锁版本号"
          },
          "stages": {
            "minItems": 8,
            "maxItems": 8,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SopStageTemplateInputDto"
            }
          }
        },
        "required": [
          "version",
          "stages"
        ]
      },
      "PublishSopVersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1,
            "description": "乐观锁版本号"
          }
        },
        "required": [
          "version"
        ]
      },
      "PreviewSopMaterialBackfillDto": {
        "type": "object",
        "properties": {
          "studentIds": {
            "maxItems": 500,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "ApplySopMaterialBackfillDto": {
        "type": "object",
        "properties": {
          "studentIds": {
            "maxItems": 500,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "previewFingerprint": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64,
            "description": "预览接口返回的内容指纹"
          }
        },
        "required": [
          "previewFingerprint"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ApiEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "nullable": true
          },
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ],
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}